firmware: move clkin_ctrl_set to clkin.h|c
